refactor: replace securityUtils with new api module for server status handling
