+++ src/core/status/status.ts
@@ -2,7 +2,7 @@ import * as net from "net";
 import { logInfo, logError } from "../../utils/logger";
 import { configManager } from "../../config/config";
 import { UserConfig } from "../../types/types";
-import { sendServerStatus } from "./utils/serverReq";
+import { sendServerStatus } from "./utils/api";
 import { ServerStatusBody } from "../../types/types";
